refactored image_classifier_head to classifier_head
